Move database migration to a premium service
